requirements.txt
Flask==3.0.0
flask-cors==4.0.0
azure-storage-blob==12.19.0
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0

